Do not parse lambda lists so many times during IR1 conversion
The following functions now have a sibling function named %NAME that
accepts (some of) the values of PARSE-LAMBDALIKE as arguments (in
particular an already parsed lambda list):

* NAME-LAMBDALIKE
* IR1-CONVERT-LAMBDA[LIKE]
* IR1-CONVERT-INLINE-LAMBDA
* FTYPE-FORM-LAMBDA-LIST
* GET-DEFINED-FUN
* STRIP-LAMBDA-LIST

Callers that call more than one of the above on the same lambda
expression can now call PARSE-LAMBDALIKE once and call the %-versions
with pre-parsed arguments.

In addition, MAKE-LAMBDA-VARS has been incompatibly changed to
accept a pre-parsed lambda list.